adds safeguard to prevent multiple firings

+++ app.js
@@ -243,19 +243,22 @@ for (let i = 0; i < allowedServers.length; i++) {
 
     client.on(Events.MessageCreate, async (message) => {
       console.log("message create");
-        // Ignore messages from bots
-        if (message.author.bot || !message.member.permissions.has(PermissionsBitField.Flags.ManageMessages)) return;
-        console.log("allowed to see this message: ", message.author.globalName, message.content);
-        // Check if the message content matches the target string
-        if (message.content.substring(0, 5) === "Grim:") {
-          const joshuaMessage = message.content.slice(5);
-          try {
-            await message.channel.send(joshuaMessage);
-            await message.delete();
-          } catch (err) {
-            console.error("Failed to send/delete message:", err);
-          }
+      if (message.guildId !== allowedServers[i]) {
+        return;
+      }
+      // Ignore messages from bots
+      if (message.author.bot || !message.member.permissions.has(PermissionsBitField.Flags.ManageMessages)) return;
+      console.log("allowed to see this message: ", message.author.globalName, message.content);
+      // Check if the message content matches the target string
+      if (message.content.substring(0, 5) === "Grim:") {
+        const joshuaMessage = message.content.slice(5);
+        try {
+          await message.channel.send(joshuaMessage);
+          await message.delete();
+        } catch (err) {
+          console.error("Failed to send/delete message:", err);
         }
+      }
     });
 
     //add emoji-related commands here. choppingblock censors posts by deleting the original and reposting it with spoiler
